+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Stage
 * Adiciona possibilidade de listar colaboradores, veículos e pneus por múltiplas unidades (PL-2695)
 * Adiciona custo de seviços no relatório de dados gerais de movimentação (PL-2733)
 
+### Features
+* Implementa tratamento de erros para o import dos mapas (PL-2410)
+
+
 Change Log
 ==========